# openapi.yaml
openapi: 3.1.1
info:
  title: Open-Spot API
  description: |
    개인 장소 기록 및 관리 서비스 API

    ## 주요 기능
    - Google OAuth2 인증 (Bearer Token + HttpOnly Cookie)
    - 개인 장소 기록 관리 (평점, 리뷰, 태그)
    - 장소 그룹 관리
    - 푸시 알림 관리

    ## 인증 방식
    - **Access Token**: Authorization 헤더로 전달 (Bearer 방식)
    - **Refresh Token**: HttpOnly Cookie로 전달 (XSS 방어)

    ## MSA 아키텍처
    - **Gateway (8080)**: 모든 API의 진입점, JWT 검증 수행
    - **Auth Service (8081)**: Google OAuth2 로그인, JWT 발급
    - **Location Service (8082)**: 장소 관리
    - **Notification Service (8083)**: 알림 관리

    **중요**: 프론트엔드는 반드시 Gateway(8080)를 통해 API를 호출해야 합니다.
  version: 1.0.0
  contact:
    name: Open-Spot Team
servers:
  - url: http://localhost:8080
    description: Gateway (모든 API 호출은 이 URL 사용)
  - url: http://localhost:8081
    description: Auth Service (직접 호출 금지, Gateway 경유 필수)
  - url: http://localhost:8082
    description: Location Service (직접 호출 금지, Gateway 경유 필수)
  - url: http://localhost:8083
    description: Notification Service (직접 호출 금지, Gateway 경유 필수)
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/locations:
    get:
      tags:
        - Locations
      summary: 장소 목록 조회
      description: 다양한 조건으로 장소를 검색합니다.
      operationId: getLocations
      parameters:
        - $ref: '#/components/parameters/NorthEastLatParam'
        - $ref: '#/components/parameters/NorthEastLonParam'
        - $ref: '#/components/parameters/SouthWestLatParam'
        - $ref: '#/components/parameters/SouthWestLonParam'
        - $ref: '#/components/parameters/LatitudeParam'
        - $ref: '#/components/parameters/LongitudeParam'
        - $ref: '#/components/parameters/RadiusMetersParam'
        - $ref: '#/components/parameters/CategoryIdParam'
        - $ref: '#/components/parameters/KeywordParam'
        - $ref: '#/components/parameters/TargetUserIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 장소 목록 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LocationSummaryPageResponse'

    post:
      tags:
        - Locations
      summary: 새 장소 생성
      description: 새로운 개인 장소를 등록합니다.
      operationId: createLocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationRequest'
      responses:
        '201':
          description: 장소 생성 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LocationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/locations/{locationId}:
    get:
      tags:
        - Locations
      summary: 장소 상세 조회
      description: 특정 장소의 상세 정보를 조회합니다.
      operationId: getLocationById
      parameters:
        - $ref: '#/components/parameters/LocationIdParam'
      responses:
        '200':
          description: 장소 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LocationResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Locations
      summary: 장소 기본 정보 수정
      description: 장소의 기본 정보를 수정합니다.
      operationId: updateLocation
      parameters:
        - $ref: '#/components/parameters/LocationIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLocationRequest'
      responses:
        '200':
          description: 장소 수정 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LocationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Locations
      summary: 장소 비활성화
      description: 장소를 비활성화합니다 (논리적 삭제).
      operationId: deleteLocation
      parameters:
        - $ref: '#/components/parameters/LocationIdParam'
      responses:
        '200':
          description: 장소 비활성화 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          locationId:
                            type: string
                            format: uuid
                          message:
                            type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/locations/{locationId}/evaluation:
    put:
      tags:
        - Locations
      summary: 개인 평가 정보 수정
      description: 장소에 대한 개인 평점, 리뷰, 태그를 수정합니다.
      operationId: updateLocationEvaluation
      parameters:
        - $ref: '#/components/parameters/LocationIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLocationEvaluationRequest'
      responses:
        '200':
          description: 평가 정보 수정 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LocationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/locations/{locationId}/group:
    put:
      tags:
        - Locations
      summary: 장소 그룹 변경
      description: 장소가 속한 그룹을 변경합니다.
      operationId: changeLocationGroup
      parameters:
        - $ref: '#/components/parameters/LocationIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeLocationGroupRequest'
      responses:
        '200':
          description: 그룹 변경 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LocationResponse'
  /api/v1/locations/{locationId}/coordinates:
    put:
      tags:
        - Locations
      summary: 장소 좌표 수정
      description: 장소의 GPS 좌표를 수정합니다.
      operationId: updateLocationCoordinates
      parameters:
        - $ref: '#/components/parameters/LocationIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLocationCoordinatesRequest'
      responses:
        '200':
          description: 좌표 수정 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LocationResponse'
  /api/v1/locations/top-rated:
    get:
      tags:
        - Locations
      summary: 최고 평점 장소 목록
      description: 내 개인 평점 기준 최고 평점 장소 목록을 조회합니다.
      operationId: getTopRatedLocations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 최고 평점 장소 목록 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LocationSummaryPageResponse'
  /api/v1/locations/recent:
    get:
      tags:
        - Locations
      summary: 최근 등록 장소 목록
      description: 최근 등록한 내 장소 목록을 조회합니다.
      operationId: getRecentLocations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 최근 등록 장소 목록 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LocationSummaryPageResponse'
  /api/v1/locations/self:
    get:
      tags:
        - Locations
      summary: 내 장소 목록
      description: 내가 생성한 모든 장소 목록을 조회합니다.
      operationId: getMyLocations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 내 장소 목록 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LocationSummaryPageResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Access Token을 Authorization 헤더로 전달합니다.

        **사용법**: `Authorization: Bearer <access-token>`

        **토큰 획득**:
        1. Google 로그인 성공 시 URL 쿼리 파라미터로 전달
        2. 토큰 갱신 API 응답 본문에서 획득

        **인증 플로우 (MSA 표준)**:
        1. 클라이언트 → Gateway: `Authorization: Bearer <token>` 헤더 포함
        2. Gateway: JWT 검증 (HS256 알고리즘)
        3. Gateway: 검증 성공 시 `X-User-Id` 헤더 추가
        4. 내부 서비스: `X-User-Id` 헤더로 사용자 식별 (JWT 검증 불필요)
  parameters:
...
    LocationIdParam:
      name: locationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CategoryIdParam:
      name: categoryId
      in: query
      required: false
      schema:
        type: string
        format: uuid
    NorthEastLatParam:
      name: northEastLat
      in: query
      required: false
      schema:
        type: number
        format: double
        minimum: -90
        maximum: 90
    NorthEastLonParam:
      name: northEastLon
      in: query
      required: false
      schema:
        type: number
        format: double
        minimum: -180
        maximum: 180
    SouthWestLatParam:
      name: southWestLat
      in: query
      required: false
      schema:
        type: number
        format: double
        minimum: -90
        maximum: 90
    SouthWestLonParam:
      name: southWestLon
      in: query
      required: false
      schema:
        type: number
        format: double
        minimum: -180
        maximum: 180
    LatitudeParam:
      name: latitude
      in: query
      required: false
      schema:
        type: number
        format: double
        minimum: -90
        maximum: 90
    LongitudeParam:
      name: longitude
      in: query
      required: false
      schema:
        type: number
        format: double
        minimum: -180
        maximum: 180
    RadiusMetersParam:
      name: radiusMeters
      in: query
      required: false
      schema:
        type: number
        format: double
        minimum: 100
        maximum: 50000
    KeywordParam:
      name: keyword
      in: query
      required: false
      schema:
        type: string
        maxLength: 100
    TargetUserIdParam:
      name: targetUserId
      in: query
      required: false
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        data:
          type: object
          nullable: true
        error:
          type: object
          nullable: true
...
    ChangeLocationGroupRequest:
      type: object
      properties:
        groupId:
          type: string
          format: uuid
          nullable: true
...
    LocationSummaryPageResponse:
      type: object
      required:
        - content
        - page
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/LocationSummaryResponse'
        page:
          $ref: '#/components/schemas/PageInfo'

    # ========================================
    # Location Group Schemas
    # ========================================
...